add kafka & rest emitters
