+++ metadata-ingestion/examples/recipes/kafka_to_datahub_rest.yml
@@ -2,9 +2,10 @@
 source:
   type: "kafka"
   config:
-    connection.bootstrap: "broker:9092"
+    connection:
+      bootstrap: "broker:9092"
 
 sink:
   type: "datahub-rest"
   config:
-    server: 'http://localhost:8080'
+    server: "http://localhost:8080"
